extract and share types to reduce repetition
